requirements.txt
# ESI-Lite Requirements
# Python 3.9+ recommended, 3.11 tested

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests for data fetching
requests>=2.31.0
chardet<6

# Configuration
pyyaml>=6.0.0

# Visualization
matplotlib>=3.7.0

# PDF generation
fpdf2>=2.7.0

# HTML generation (for email templates)
jinja2>=3.1.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Optional: For enhanced data fetching
# (Not required for basic functionality)
# beautifulsoup4>=4.12.0  # For HTML parsing if needed
# lxml>=4.9.0             # For XML parsing if needed

# Development dependencies (optional)
# pytest>=7.4.0           # For testing
# black>=23.0.0           # For code formatting
# flake8>=6.0.0           # For linting
# mypy>=1.5.0             # For type checking
